_init/cloud_function/alerting/requirements.txt
google-cloud-bigquery>=3.0.0
duckdb>=0.10.0
pandas>=1.5.0
pyarrow>=10.0.0
